Adding missing __setitem__ and keys() functions
Because the __enter__ function returns the implementation,
using

with DataFile(..) as f:
  ...
  f["somevar"] = a

would not work; the implementations did not have __setitem__ methods.

The DataFile class did not have a keys() function, though the
implementations did. This makes them consistent by adding a keys()
function to DataFile. This does the same as list(), but is more
consistent with the python standard containers.