Calculate correct guard cell values in multigrid Laplace solver
Multigrid solver does not correctly enter the blocks that should set the
guard cells of the result (multigrid_laplace.cxx:385 and 403). This
commit changes the tests for boundary processors to mesh->firstX() and
mesh->lastX(), which is consistent with how the boundary conditions are
input from x0.

The boundary conditions were also not correctly applied in the multigrid
laplacian solver to give the guard cells.