Catch errors in interp_to with Field2D
const Field2D interp_to(const Field2D &var, CELL_LOC loc) does not
actually implement any interpolation. Previously, it just returned var
unaltered. In this commit, var is returned in cases where no
interpolation is needed (loc==CELL_DEFAULT, location of var and loc are
the same, or the interpolation would be in the z-direction [where
nothing needs to be done since a Field2D is axi-symmetric]); otherwise
an exception is thrown.