Adding restart.addvar routine for Python
Adds a variable with constant value to all restart files.
This is useful for restarting simulations whilst turning on
new equations. By default BOUT++ throws an error if
an evolving variable is not in the restart file. By setting an option
the variable can be set to zero. This allows it to start with
a non-zero value.