Adding -l,--log command-line argument
Allows the name of the log files to be changed. The default is
set to "BOUT.log" in the DEFAULT_LOG variable. This can be changed,
for example

    --log  log

would produce log files in the input/output directory ("data" by
default) called "log.0", "log.1" etc.

This fixed issue #745

Includes an integrated test which checks -l, --log and -d options.
The test is quick, so added to the test suite.