Add interpolation to last boundary point for staggered fields
The final grid point, which is on the boundary, is xend+1 for CELL_XLOW
fields and yend+1 for CELL_YLOW fields, if the domain is not periodic.
Since there are 2 CELL_CENTRE guard cells beyond this point, interp_to
can calculate the value there. This commit enables this by introducing a
new REGION, RGN_INTERP, which is then used for the loop in interp_to.

Updated staggered_grids.rst section of manual to note this new
behaviour.