Introducing thread-arena based array.
This essentially provides a separate data store for each thread. This
allows us to remove all openmp pragmas from this implementation of Array
and removes any synchronisation issues that were here.

This doesn't appear to change the runtime very much in my test case but
will hopefully be useful in general.