Reduce the number of synchronisations in the openmp fft routines
Original approach  had a blocking synchronisation on *every* call. This
block typically didn't need executing.

Here we've made two main changes

1. Replace critical with single as this more accurately represents what
we want to do.
2. Swap the pragma and `if` lines. Originally it was blocking pragma and
then `if` block that usually was not executed. Now as the `if` statement
comes first and the sync is inside this block we only sync on the rare
occassions when the `if` block is executed.

This saved 20% from total run time for a sample problem with openmp.