Python tools: caching and passing kwargs to collect in BoutOutputs
Implement caching of variables in BoutOutputs. Caching is optional, in
case memory is limited, and defaults to False to replicate the current
behaviour. Also added as option to BoutData, which passes the option to
BoutOutputs.

Add **kwargs argument to BoutOutputs and BoutData which is passed to
collect when any variable is collected, so that the user can change
collect's options when using BoutOutputs.