Don't check for C++ when configuring PETSc
PETSc documentation
(https://www.mcs.anl.gov/petsc/documentation/installation.html) says
"Note that --with-clanguage=cxx means that the PETSc source code is
compiled with the C++ compiler. This is not normally needed and we don't
recommend it. One can use 'c' build of PETSc from both C and C++." So
using the C build should be fine and we don't need to check that PETSc
was built with a C++ compiler.