pylib/boututils/showdata.py: x,y positions in showdata
Allow single arrays to be passed in for the x and y arguments of
showdata. These will then be used for all variables in the vars
argument. Also add transposes in call to contourf so that the shapes of
x and y match that of the data (which was transposed already).

This update allows, for example, poloidal planes to be plotted in
physical R-Z coordinates.