Add unit test for invalidateGuards and fix implementation of routine.
Note also had to disable this check (by only allowing if CHECK>4, which
can only be achieved by manual edits of makefile). This is because as it
stands invalidating the guard cells causes several integrated tests to
fail presumably due to accessing values in the guard cells that have not
been set. These may be indications of errors, but likely in some places
we're just using the default region (i.e. RGN_ALL) when we only really
need the no boundary version.