Use checkData() instead of ASSERT3(finite(...))
Replace use of ASSERT3(finite(...)) in loops in functions in field3d.cxx
and field2d.cxx with checkData(result) after the loop. This means that
the guard cells don't get checked, which resolves the issue of checks
failing when operating on fields whose guard cells are allowed to be
invalid (e.g. results of differential operators).