Add wrappers for Python warnings
Adds boututils.boutwarnings module containing alwayswarn() and
defaultwarn() functions.

alwayswarn(message) raises an instance of AlwaysWarning, which is set by
default to always be printed. This replaces the default behaviour of
UserWarning, which is to be printed only once, the first time a
particular warning call is encountered. This is useful, e.g., for a
function that may be called interactively where the warning is about the
consequences of the inputs to the function.

defaultwarn(message) is just a convenience wrapper for
warnings.warn(message).

The new alwayswarn() is now used in showdata().