Add test of multigrid Laplacian solver
The test uses a sheared grid to test more of the terms than an unsheared
grid would.

Test uses 1 and 3 MPI processes. 3 is the minimum number to check all
types of configuration of presence/absence of boundaries on a processor.

Metrics in the BOUT*.inp files were copied from some binary grid files.
y-dependence of metric terms is introduced to make G3 non-zero.

Note that the sheared grid tests use a fairly high atol=1.e-10 because
the RHS in those tests comes out fairly large, O(10^6).