Fix intermittent bug in multigrid solver
A check for slow convergence of GMRES iteration was allowed to stop the
outer iteratitive loop of the multigrid solver if the error did not
decrease fast enough, even though the tolerance check against rtol and
atol had not been passed. This should only be allowed in inner solvers
(in particular the GMRES solver at the coarsest level) where if it
introduces an error it just means that the outer iterative solve has to
continue.