Correct checks for number of multigrid levels
In the multigrid solver, the number of grid points in both directions
must be divisible by 2^(mglevel-1) where 'mglevel' is the number of
multigrid levels. Also 'aclevel', the number of levels before collecting
to a serial multigrid solver, must satisfy the condition that
2^(aclevel-1) divides the local number of grid points on each processor,
in both directions.

Previously, only aclevel was checked. This commit adds a check on the
global mglevel before the check on aclevel.