Suppress unnecessary warning
When there is only one grid point on the coarsest level of the multigrid
solver, one of the 'a1' values becomes zero. This is checked for because
a0=1/a1 is used to calculate values for the next iteration, but when
there is only one grid point, the solution is exact up to floating point
error and the residual will turn out to be zero, so that we exit the
loop on the first iteration. Therefore in this case the the 1/0 value
will never be used, and the calculation can continue correctly so the
warning is not needed. This commit suppresses it when there is one point
on the coarsest level.