Remove `(*this) = f` in Field2D copy constructor
This duplicates the work done in the copy constructor but without some
of the checks that take place and can end up clobbering values (e.g. we
may set `nx=-1` in some situations in the copy constructor, which are
then just copied from the input field in the `operator()` call).