Allow runtime control of maximum size of contiguous block to use in Region
Adds `mesh:maxregionblocksize` input option. Defaults to whatever is
defined by MAXREGIONBLOCKSIZE (i.e. defaults to 64).

Also adds an example in examples/performance/tuning_regionblocksize that
shows timing information for `r = a + b` using a BLOCK_REGION_LOOP and
using different block sizes.

This could be used as a way of picking a good value of the default
block size for a particular machine/configuration.