Fix petsc_laplace bug and remove global mesh
'FieldPerp sol' was initialised without any mesh, and caused the Laplace
solver to fail.

Removed all references to the global 'mesh' variable, replacing with
localmesh, which can be set with an argument to the constructor,
defaulting to global mesh. The solve() method checks that the rhs passed
in has the same mesh as the solver object.