Extend #481 slightly to be more rigorous in checking
Rather than just checking if the diff type is known, also check if it is
valid for the current differencing type (first, second, upwind etc.).

Specifically this will now catch the case where `mesh:ddz:upwind=fft`,
previously this would pass the checks as `fft` is a known type, but
because it's not implemented for `upwind` this would actually (silently)
revert to `u1`.