Optimize collect in BoutOutputs
collect() spends a lot of its time opening the DataFile objects. In
BoutOutputs it makes sense to make a cache of all the DataFiles so they
only need to be opened once. This speeds up subsequent collects (I've
seen about 5x speedup).

Implemented using a method of BoutOutputs that mostly duplicates the
collect function, except for using a cache for the DataFiles.