Optimize collect in BoutOutputs
collect() spends a lot of its time opening the DataFile objects. In
BoutOutputs it makes sense to make a cache of all the DataFiles so they
only need to be opened once. This speeds up subsequent collects (I've
seen about 5x speedup).

Adds a function create_cache to boutdata.collect that creates a list of
DataFiles, returning 'file_list', 'parallel', 'suffix' and
'datafile_list' as a namedtuple. This namedtuple can be passed as an
optional argument 'datafile_cache' to collect.

In BoutOutputs, by default a cache of DataFiles is created the first
time collect is called, and then passed to collect each time. This can
be switched off by passing DataFileCaching=False to __init__.