Fix bug in D2DY2 with staggergrids=true and outloc=CELL_DEFAULT
The interp_to in the return statement of D2DY2 is redundant and can be
removed. It is a bug, because if D2DY2 is called without an outloc
argument, CELL_DEFAULT is passed as the location to interp_to().
interp_to does not handle CELL_DEFAULT, so an exception is raised unless
staggergrids=false so that interp_to does nothing.