Disable unimplemented staggered derivatives, replace with checks
Some staggered derivatives, in particular between two different
staggered grids, previously attempted to operate by interpolating and
then taking derivatives, or taking derivatives and then interpolating.
Neither of these work, because after a derivative/interpolation the
field needs to be communicated and have boundary conditions applied
in order to set the guard cells before another interpolation/derivative
can be taken. Since this behaviour is not ever used, it is better to
throw an exception if it is somehow requested, and to check the final
location of the result instead of calling 'interp_to' at or just before
the return statement.