Field cell location saved to dmp file
The datafile class now sets an attribute "cell_location"
for each 2D and 3D field (and vector component).

The attribute is a string which is the same as the enum name
e.g. CELL_CENTRE, CELL_YLOW etc.

To do this, a const map has been added to bout_types.hxx, together
with an inline function CELL_LOC_STRING(). This replaces
`strLocation` in interpolation.hxx, which used a switch.

Together with PR #770 (saving attributes in NetCDF)
and #996 (reads attributes into a BoutArray class), this now
works (in test-io):

```
> f3d = collect("f3d")
> f3d.attributes
{'cell_location': 'CELL_CENTRE', 'bout_type': 'Field3D'}
```

Added some checks to `test-io` to check that the cell_location
attribute is read correctly.