Add missing fromFieldAligned() in interp_to()
In interp_to() when yup/ydown fields are not present, we use
toFieldAligned() to transform to field aligned coordinates before
interpolating in the y-direction, but didn't transform the result back
to the original coordinates.

This does not cause an error when the ParallelTransform is
ParallelTransformIdentity, because toFieldAligned/fromFieldAligned do
nothing in this case, and this is probably where interp_to has been used
so far.