Add separate option for staggered derivatives
If StaggerGrids=true, then only methods implemented for staggered
derivatives could be used for unstaggered derivatives, e.g. FFT
derivatives could not be used for unstaggered derivatives. This commit
adds an option which, if set explicitly, allows a different method to be
chosen for staggered derivatives to avoid the exception being thrown for
unknown option when it is only unknown for the staggered case. If this
option is used, presumably the staggered derivative should not be used,
but since the user must set the option explicitly, they must be aware
that different methods would be used.