Finish implementation of Flexible<>
Don't delete member fields of a Flexible<>: writing Flexible fields to
output sometimes failed because the reference to the field that had been
passed to the output file became invalid. To fix this (i) implement
passing variables to Datafile by pointer and (ii) don't delete the field
pointers in a Flexible, instead if a field needs to be changed, just
assign to the existing field. Replace 'clean' which deleted fields with
'reset' which declares them invalid (and sets to nan if CHECK>2).

Define copy constructor and copy assignment operator so they are not
defined explicitly and used (this causes memory problems because they
would copy pointers).

Initialise xlow and ylow fields of Bxy in Coordinates, including guard
cells (by communicating) and boundary guard cells (copied from nearest
grid cell). This is needed because Bxy is used in arguments to
derivatives in some differential operators.

Add in-place operators ('+=(Flexible<F>', etc.) for Field3D and Field2D.
Note, need forward-declaration of Flexible class in field3d.hxx and
field2d.hxx to avoid problems due to circular includes.

Add default values for all fields in Coordinates constructor. Otherwise
we get execptions throw by checkData when a temporary Field2D is copied
into the Flexible<Field2D>.

Use 'set()' explicitly to set non-mainid locations of a Flexible field.
Previously assignment operator was used, but this now resets the
Flexible, using the rhs Field as the 'main' field. Update wave1d test to
do this.

In the 'FakeField' of tests/unit/field/test_flexible.cxx: add default
value for constructor argument; implement isAllocated(),
operator[](const DataIterator &), iterator(), begin() and end() to allow
FakeField to be used an an 'auto' loop, used when CHECK>2 by
Flexible::reset().

Only get references to const fields from Flexible: remove getNonConst.
Instead, when adding a Flexible field to a Datafile, use a const_cast<>
to get a reference-to-non-const for 'main' field (fields[mainid]).
With a reference to const we cannot read from files.