Remove mixed field-aligned/non-field-aligned derivatives
Advective and flux derivatives, which take v and f as inputs, had cases
using yup/ydown fields for one of v and f, but not the other. These do
not make sense as multiplication of a field in field-aligned coordinates
with another in non-field-aligned coordinates is incorrect.

Delete last of these cases and fall back to converting both v and f to
field-aligned if either does not have yup/ydown fields.

Also add some comments where mixed cases were removed before.