Allow for outloc=CELL_DEFAULT when using Flexible::get()
In many derivatives, we were calling, e.g. SQ(g_11.get(outloc)) where
outloc might be CELL_DEFAULT. In this case, get will return the 'main'
field (which will be CELL_CENTRE for the geometrical coefficients) but
this may not match the location of the field. We can fix this problem by
using the location of the result of the call to indexDD*, etc., to set
the location.