Fix use of unassigned values in ShiftedMetric
To ensure corner guard cells are not read, restrict all loops in
ShiftedMetric to not include x-guard cells.

Add region arguments (where it is valid to give RGN_NOX or RGN_NOBNDRY)
so functions like toFieldAligned/fromFieldAligned can be used both when
y-guard cells have been set and when they have not. e.g. when using
fromFieldAligned on the results of interpolation or derivatives, y-guard
cells cannot be included, but when using toFieldAligned on the inputs to
interpolation or derivatives, y-guard cells must be included.

Also add some extra calls to invalidateGuards to help catch future
errors.