Fix mms_alternate.py
Major change is to correct use of 'x' as poloidal flux (starting from
inner edge of grid) vs. [0,1].

Fix calculation of dx/dy/dz. Were previously, incorrectly, using Lx as
a box length in all directions. Should be grid width of psiwidth in
x-direction and 2pi in y- and z-directions.

Add psiN0 argument of SimpleTokamak to choose where in minor radius
to start the grid.