Allow t_array to be passed to showdata()
t_array is used to show the simulation time in the title of movies.
Previously it was taken from 'collect('t_array')' which assumes that
showdata is run from the same directory as the data. This commit adds an
optional t_array argument to showdata that can be used to pass in the
time, with default value 'None' keeping the old behaviour.