Use slice objects for arguments to DataFile and collect
DataFile.read() takes a ranges argument. This was previously a list of
integers of size 2*ndims, now is a list of slice objects of length
ndims. Alternatively, instead of a slice a list or tuple of length
between 1 and 3 can be given which is converted to a slice. Or an int
can be given which is converted to a slice that picks one point (a
list/tuple of length 1 is treated as an int).

collect() takes tind/xind/yind/zind arguments, which were previously
ints or lists of length 2. Now they are slices, the same as can be
passed to DataFile.read()'s ranges argument.