Guard against double-frees of PETSc objects
PETSc registers some objects on creation, and will destroy them itself
on calls to PetscFinalize. If an object survives past our call to
BoutFinalise, which calls PetscFinalize unconditionally, we may get
double-frees when we try to destroy PETSc objects ourselves.

Solution is to check if PetscFinalize has already been called, and if
so, don't destroy those objects.