Add monotonic Hermite spline interpolator
Cubic splines have high accuracy, but necessarily produce overshoots
so that the maximum value of the interpolating function can be higher
than any of the nodes. This can cause problems, for example with
Neumann boundary conditions a diffusion equation can still grow (and
become unstable) due to these overshoots.

This adds MonotonicHermiteSpline, which is almost the same as
HermiteSpline, and so inherits from that class. This new interpolator
adds a check that the interpolated value does not go outside the
min/max range of the surrounding grid points.

This will reduce accuracy near maxima and minima, but elmininates the
overshoot oscillations.