Set guard cells after taking derivatives
In LaplaceNaulin, the guard cells of the field 'x' must be copied from
the guard cells of the input 'x0' before calling solve, in order to pass
boundary values to 'delp2solver' if required. Previously this was done
before calculating x-derivatives of the field 'x', potentially causing
an error next to the boundary.