Fix arithmetic in mean(Field3D) and tidy up code
In the case when allpe=true, mean(Field3D) would previously give the
wrong result. To fix this, calculate and MPI_Allreduce (using MPI_SUM)
both the sum of the field values and the number of grid points. So that
this can be done in a single communication, store both values in a
2-element array.

Calculate npoints in the loop instead of in a switch-case before it to
make the code more flexible when/if more regions are added.

Correct some comments which referred to 'max' instead of 'mean'.