The simple version does not work.
Take for example 10 and 1e-4 - obviously they should be multiples.
The first fmod returns 1e-4 - as expected.
The second however returns 0.00099999999999979185 - as 10 is slightly
smaller than 1e5 * 1e-4 - due to the finite accuracy of floating point
numbers.
Even if we do not compare for zero - but a finite small value, the
result will be negative, as 10 was slightly smaller - compared to the
divisor - and thus the result of fmod is in the order of the smaller
number.