Fix showdata error when plotting 0.
If a variable is 0. showdata's 2d plots previously failed because
contour levels are set explicitly and matplotlib throws an exception if
they are not different. This was fixed for most cases by adding an
offset proportional to the value of the field being plotted, but a field
which is everywhere zero still caused an error. This commit adds a very
small, 1.e-36, absolute offset to fix this.

Also, while fixing the global_colors=True branch, tidy up the loop which
was unnecessarily setting a min/max for each field equal to the global
min/max and then checking them.