Set location of result as soon as it is declared in index_derivs.cxx
The location of the result will in future need to be set before shifting
it from field-aligned coordinates, if this is necessary. So it is safer
to set the location to outloc as soon as 'Field3D result' is declared.

Also in a couple of places just 'return apply*diff(...)' instead of
creating an unneeded intermediate variable 'result'.