Set location of result as soon as it is declared in index_derivs.cxx
The location of the result needs to be set before shifting it from
field-aligned coordinates, if this is necessary. So it is safer to set
the location to diffloc (or outloc where diffloc is not yet set) as soon
as 'Field3D result' is declared.
There are a couple of  exceptions where the location is set after
calling apply*diff (which should not be necessary, but does not hurt,
except that it may hide errors setting location in apply*diff) where the
location of 'result' would not be used anyway earlier in the method.