Escape variable names when searching expressions
Bugfix for expression parsing in BoutOptions/BoutOptionsFile.
Otherwise special characters are interpreted as part of the regex. This
causes a problem in particular when reading BOUT.settings files: if the
command line option '-d .' is used, then an option '. = True' ends up in
BOUT.settings and the '.' was interpreted as a search pattern.