Write attributes to file only once
Don't write attributes every time variables are saved to an output file.
Instead, (i) attributes are written to the file when they are added, and
(ii) the location of Field2D and Field3D variables is written as an
attribute only the first time Datafile::write() is called.

For (i), variables have to be added to the file so that attributes can
be added to them, even if they have not been written yet. However, they
cannot be added if the file has not been opened for writing yet.
Therfore: added a new flag 'writable' to the Datafile class, which is
set to true by openw() and opena() and to false by openr() and close();
when a variable is added, if writable==true then the variable is created
in the file, otherwise it is just added to the *_arr vector, and is
created in the file when openw() or opena() is called.

For (ii) a member 'first_time' is added to Datafile and the locations
are set as attributes when first_time==true. first_time is set to true
when Datafile is created, and when openw() or opena() are called. It is
used to detect when Datafile::write() is called for the first time.