Search for expressions as whole words when evaluating options
When searching an expression for a variable name, surround the variable
name with '\b' so that only whole words match. This prevents infinite
recursion that could previously occur when a variable name matched a
substring of an expression, but wasn't actually that variable. E.g. if
the option is 'omega = omega_ci*delta' evaluate_scalar() might try to
evaluate omega, find omega as a substring of omega_ci, then try to
evaluate omega again, etc.