Make Monitor::timestep protected rather than private
Allows subclasses to overwrite the value of timestep in their
constructors, which is useful for setting its value from options in
BOUT.inp. When timestep was private its value could only be set in the
initializer list, when the parent constructor Monitor::Monitor() was
called, which is much more restrictive.