Use `size_type` to define the type of size related variables in Matrix and Tensor
Currently set this to `int` whilst previous behaviour would be
equivalent to `unsigned int`. The motivation for changing to `int` is
that this allows `operator()` uses to vectorise when in loops with an
`int` index.

Changing loop indices to `unsigned int` (or `std::size_t`) doesn't
actually allow vectorisation even when `Matrix::size_type = std::size_t`
as `Array` currently also uses `int` as the indexing type. This seems to
be enough to prevent vectorisation in simple loops over `Matrix` elements.