Check for None explicitly in collect
When checking/converting tind/xind/yind/zind arguments to collect, was
previously using 'if not r' to check if 'r' is None, but this gives true
also if r is an integer 0. r=0 should be treated differently from r=None
(should go through to the integer case). Therefore need to use 'if r is
not None' explicitly as the test for None.