Replace FFT approach to use BOUT_FOR
Note this assumes there is a valid Region2D for the region name passed
in. For this reason there is currently an assert that the requested
region is one of the known default types.

Implements fft filtering for the 2nd order derivative.

Removes assert to check that nz is a power of 2.