Solver option loadsave_chunks for faster data transfer
New option `solver:loadsave_chunks` modifies how data is copied
between BOUT++ Field objects and the 1D arrays used internally in the
time integration solvers.

The old method (= `false`, the default) interleaves the fields, so that values
which are spatially close together (and so likely to be strongly
coupled) are close in the 1D arrays. This is used in some
preconditioners, e.g. block diagonal, which assume nearby points are
more likely to be coupled.

The new method (= `true`) puts each field into a contiguous
block. This is faster, but now the coupling will not be near the
diagonal. In many cases this may not matter.