PhysicsModel: use Monitor subclass for monitor
Previously PhysicsModel defined an outputMonitor method that was
specially called by the Solver. This meant that the method was called at
the frequency of the fastest added Monitor.

Instead, make a friend class PhysicsModelMonitor that is a subclass of
Monitor and contains a pointer to the PhysicsModel object. This can be
added using Solver::addMonitor() and so is called at the output
frequency by default.

The PhysicsModelMonitor is added to the solver in postInit, so the user
can call modelMonitor.setTimestep() in init() to change the output
frequency, if desired.