Don't use setLocation for setting default field location as already set (in header)
This is to try to avoid segfaults experienced in Travis testing.

Note this means that `fieldCoordinates` will be nullptr on
initialisation and with most constructors. Could try to address this in
setLocation and have added a comment to this effect in the appropriate
routines, however this currently leads to unit tests that
crash (presumably due to operation with an inappropriately setup mesh).