Initialize skip_mask with local mesh
In Interpolation, avoid skip_mask being created with the global mesh if
a local mesh is present. Instead of using 'Mesh* mesh=nullptr' as the
default, use 'Mesh* mesh_in = mesh' so that localmesh is always
initialized. Then skipmask can be initialized in the initialization
list, using mesh_in, when no mask is passed in. This avoids the default
constructor 'BoutMask()' being called, bringing in the global mesh, as
it was before if the Interpolation constructor without BoutMask argument
was called.