Adding OMP parallelisation for loops in rkgeneric solver
Note this introduces some slightly non-deterministic behaviour due to
the parallelisation of the calculation of the error estimate. Now
floating point operations may be in a slightly different order each time
and hence give slightly different round-off. This could lead to a small
variation in the number of right hand side calls between repeat runs.