Access options using dictionary interface
Modifies the internal storage of Options, adding new interface
to values and sections. Backwards compatible with previous interface.

An Options object is now a union type, which can be either a section
or a value, depending on how it's treated. This is enforced, so once
an Option is treated as a value (e.g. assigned a value) then it cannot
be treated as a section.

Replaces
```
int value;
Options::getRoot()->getSection("mysection")->get("myvalue", value, 42);
```

with
```
int value = Options::root()["mysection"]["myvalue"].get(42);
```
or
```
int value = Options::root()["mysection"]["myvalue"];
```
Since no default is given here, if no value is found then an exception
will be thrown.