Options, FieldFactory Fixes, const changes, similar function
FieldFactory now uses `const Options*` rather than `Options*`
This has no real effect on the interface, but allows it to be used
inside `Options::as<type>() const`.

The `withDefault` functions have been replaced by `as` methods,
and a `similar` template which tests if values are similar.

Add const versions of `get` and `getSection`; make `getParent` and
`isSet` const.