Modify OPTIONS macro, add pointer function
`OPTIONS(section, key, value)` is a useful macro for
making big lists of options readable. Previously this only
took `Options*` pointers for `section`, not references. This commit
enables `OPTIONS` to take either.

The following are now equivalent:
```
OPTIONS(Options::getRoot()->getSection("mesh"), nx, 10);
OPTIONS(Options::root()["mesh"], nx, 10);
```

To do this, added templated function `pointer` to `utils.hxx` which
converts either a reference or a pointer to a pointer. The reason for
converting to pointer rather than reference is to enable better error
messages in the case of `nullptr` (not handled inside `pointer`
function).