Add MACRO_FOR_EACH for variadic macros
Rather than needing a separate macro for each number of arguments
e.g. SOLVE_FOR(a), SOLVE_FOR2(a,b), SOLVE_FOR3(a,b,c) etc.
this provides a way for a macro to expand a variable number of
arguments. Currently limited to 10, but could be expanded quite
easily if needed.

The definition of `SOLVE_FOR` becomes:

```

  MACRO_FOR_EACH(SOLVE_FOR1, __VA_ARGS__)
```

For now the other macros are kept for backward compatibility.

Modest amounts of black magic are needed, but as far as I know
all standard, no compiler extensions.