Remove Field3D::mergeYupYdown()
This was used to set yup/ydown to point to the field itself. Used by
ParallelTransformIdentity. It is unnecessary because there are
field-aligned versions of all the operators, which can be used when
hasYupYdown()==false.

Removing the method simplifies several tests, which can now be just
'var.hasYupYdown()' where previously they also had to check 'var.yup()
!= var' and 'var.ydown() != var'.

Also rename 'splitYupYdown()' to 'createYupYdown()' since it no longer
makes sense to contrast 'split' with 'merge'.