Add canCalcYupYdown() methods
Allows finding out from the Mesh* whether the ParallelTransform supports
yup/ydown fields. This is used so that Field2D::hasYupYdown() returns
true if yup/ydown are supported for Field3D and false if not; derivative
operators on Field2Ds can then use the yup/ydown branch if supported for
Field3D (so only 3 point stencils are required/allowed) and the
to/fromFieldAligned branch otherwise.