Bugfix: extrapolate geometric quantities into guard cells
Previously used Neumann boundary conditions for metric components (and a
couple of other geometric quantities) to set the guard cells when
interpolating to staggered grids. This caused things that depend on
derivatives of the metric, like G1/G2/G3 to have very different, and
wrong, values at the boundaries. Now extrapolate instead, with the same
stencil as free_o3 boundary conditions, but extrapolating only the
points past the boundary. An extra interpolation has to be added after
calling interp_to to get the interpolated field at the upper/outer
boundary.