Fix extrapolation of metric when there are not enough grid cells
When metrics are calculated on staggered grids, they are extrapolated
into the boundary cells. Extrapolation requires 3 grid points. If there
are fewer than 3, just copy the nearest value into the guard cells,
since probably derivatives aren't being taken in that direction anyway.

If only 1 guard cell and 1 grid point per processor in a certain
direction are being used, then only have 2 points to extrapolate from on
a processor, so cannot extrapolate to guard cells with 3rd order
accuracy. In this case throw an exception because the user needs to
change MXG/MYG or NXPE.