interpolateAndExtrapolate: don't communicate original field
We have 'result = interp_to(f, location, RGN_NOBNDRY)' but then
communicate result. Need to ensure result's data is unique. Otherwise
result might be a duplicate of f (if no interpolation is needed, e.g. if
interpolation is in the z-direction); then f would be communicated.
Since this function is used on geometrical quantities that might not be
periodic in y even on closed field lines (due to dependence on
integrated shear), we don't want to communicate f. We will sort out
result's boundary guard cells, but not f's so we don't want to change f.