Remove mixed yup/ydown-fieldaligned cases in advective/flux derivatives
In derivative operators with two fields, it doesn't make sense to mix
fields with yup/ydown and fields shifted to a field-aligned grid,
because the grid points are in different physical locations, even the
points at the centre of the derivative stencils. Therefore remove the
mixed cases in VDDY and FDDY; now either both fields have yup/ydown or
both are shifted to field-aligned.