Add twist-shift in PF region
If twist-shift is needed in the core, it is also needed in the PF
region, but with the opposite sign.

Replace the TS_down_in, TS_down_out, TS_up_in, TS_up_out bools in
BoutMesh with ints that are allowed to be -1, 0, +1 so we can pass the
sign of the twist-shift for core vs. PF.

Set a twist shift for PF region(s) if needed.