Set branch-cut guard cells of zShift using communicate and ShiftAngle
Don't extrapolate zShift at the branch cut location, instead communicate
and then correct zShift for the discontinuity by using ShiftAngle, which
gives the jump in zShift across the branch cut.

ShiftAngle is now stored in Coordinates not just in BoutMesh. It is
interpolated (for CELL_XLOW) to the correct values for the location of
the Coordinates object.

The 'interp(stencil s)' function is moved from interpolation.cxx to
interpolation.hxx so it can be used to interpolate ShiftAngle in
Coordinates::Coordinates(CELL_LOC location).

periodicY(int jx, BoutReal &ts) now has a third argument 'CELL_LOC
location' to account for the different possible locations of ShiftAngle.
Both versions of periodicY can no longer be const, because a call to
getCoordinates(location) may require a new Coordinates object to be
created.

ShiftedMetric now requires TwistShift=false.