mms_alternate.py: make zShift continuous in PF region
Subtract shiftAngle from zShift in the lower, outer PF region to make
zShift continuous in the lower PF region (even though y is
discontinuous, y<0 in lower, inner PF and y>2pi in lower, outer PF).

Use Heaviside functions to subtract shiftAngle in the right region. This
means DiracDelta functions appear when taking derivatives. Fudge to set
these to zero because we don't actually want derivatives across the
branch cuts (we only need derivatives in the regions where the Heaviside
functions are constant), so it's fine to replace the DiracDelta's with
0.