Add Mesh* pointer to FieldData
Add a Mesh* pointer and getDataMesh() method to FieldData, so that
FieldData::setBoundary() can use the local mesh.

getDataMesh() method is needed in case the FieldData is constructed
before the global mesh is created (i.e. when Field3D, etc. are declared
in global scope). It must not be called getMesh() because that would
clash with the method of Field.