Allow escaping of characters in expressions
If variables have names which include symbols like + then expressions
need to escape the symbols to avoid treating them as arithmetic
operators. This can be important when evolving variables with names
like "H+" or "He+2".

Two mechanisms currently allowed:

1. A backslash (\) escapes the following character, so an expression
   could include something like "h2\+:density" as a symbol

2. Backquotes escape a sequence of characters, so "`h2+:density`" or
"h`2+`:density" could be used.

Note that the colon symbol always indicates section, and is not
escaped. This is to avoid ambiguity. The options reader therefore
checks if keys include `:` and throws if they do.

Unit tests added for all these things