Throw exceptions with helpful error messages.
BoundaryOp::applyAtPoint() and BoundaryOp::applyAtPointStaggered()
should never be called, but must be implemented in the base class in
case a subclass overrides BoundaryOp::apply() and so does not need
these methods. This commit provides a helpful error message if these
methods are ever called, instead of the previous ASSERT1(false).