Initialize Mesh::coords_map, Mesh::coords_xy_corner in BoutMesh::load()
Initialize Coordinates objects in BoutMesh::load() instead of waiting
for getCoordinates() to be called. This avoids issues with the
Coordinates initialization using and changing the bndry iterators, which
previously could lead to errors in calculating boundary conditions the
first time around. Also makes the Mesh::getCoordinates() and
Mesh::getXYCornerCoordinates() methods trivial, so hopefully they can be
inlined.