Only correct zShift if ShiftAngle is set, but check in ShiftedMetric
Where there is a branch cut, we correct the guard cell values of zShift
using ShiftAngle. This is not possible if ShiftAngle is empty, so skip
the correction in this case.

Check that ShiftAngle has been set in ShiftedMetric, because
ShiftedMetric relies on zShift and so will require the correction at
branch cuts.