Make hasYGuards() return false for GridFromOptions
core guard cells may not be correct if expressions are set to account
for differences (e.g. in zShift) between core and PF regions. PF is y<0,
y>2pi (for single null...) but y<0 and y>2pi in the lower/upper guard
cells of the core.

Since no Grid classes have hasYGuards true we should probably remove
this method and uses of it...